Change named params to default to None instead of False
This change prevents linters from complaining about type assignment
mismatches. The linter sees the default boolean value set in the function
definition for searchCPE() and searchCVE() and assumes the values should
be booleans, so when code sets, for example, pubStartDate and pubEndDate
as strings or datetime objects the linter complains because it is
expecting a boolean. Also, defaulting optional named parameters to None
is more common "pythonic" practice.